[apps/reg] Fix cursor navigation bug
When going UP, the cursor was sometimes stuck on a point on a regression, switching
between the point and the regression endlessly. Example data set:
X1 = [1, 2, 3, 4, 5] and Y1 = [1, 5, 6, 77, 5]

+++ apps/regression/store.h
@@ -30,11 +30,11 @@ public:
   }
   /* Return the series index of the closest regression at abscissa x, above
    * ordinate y if direction > 0, below otherwise */
-  int closestVerticalRegression(int direction, float x, float y, int currentRegressionSeries, Poincare::Context * globalContext);
+  int closestVerticalRegression(int direction, double x, double y, int currentRegressionSeries, Poincare::Context * globalContext);
   // Dots
   /* Return the closest dot to abscissa x above the regression curve if
    * direction > 0, below otherwise */
-  int closestVerticalDot(int direction, float x, float y, int currentSeries, int currentDot, int * nextSeries, Poincare::Context * globalContext);
+  int closestVerticalDot(int direction, double x, double y, int currentSeries, int currentDot, int * nextSeries, Poincare::Context * globalContext);
   /* Return the closest dot to given dot, on the right if direction > 0,
    * on the left otherwise */
   int nextDot(int series, int direction, int dot);
